Adds a route for editing the user details based on whether they are admin or user. Adds a component for editing the details.
EditDetails.jsx takes 2 arguments, userType and userID. userType determines whether they are user or admin, thus specify either one in string format else it will not work. userID is used to reference the user's ID in the database and therefore should be used by inputting the user's current session. If the admin wants to change the details of another user, then just input that user's id and specify that it is a user.